Added new ability to gather repository data
The github-activity-metrics-tool.py script was updated to include new functionality that allows for gathering of information about repositories under accounts that are predicted to be associated with a specific institution. This gathered repository information is saved as a CSV when the code is run. The code has also been made easier to adapt for other institutions by moving key parameters to the top of the script.